refactor folder structure (#92)
* refactor folder structure

* country selector navigator

* factories

* version
