+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: phone_form_field
 description: Flutter phone input integrated with flutter internationalization
-version: 4.5.3
+version: 4.6.0
 homepage: https://github.com/cedvdb/phone_form_field
 
 environment:
